feat: Prevent allocation of one regex per check (#297)
Current implementation of `check` macro for `Match` constraint allocate
and compile the regex pattern every time.
This is not a problem when using compile time constraints, but runtime
checks are slow and allocates a lot of memory because of this.

Solution is to precompile the regexp in a dedicated RuntimeConstraint
instance.